fix(temporal): satisfy CI guardrails
