+++ packages/temporal/src/activities/agent-task-runtime.ts
@@ -1,0 +1,101 @@
+import { Context } from "@temporalio/activity";
+import * as Sentry from "@sentry/bun";
+import { getTraceContext } from "#observability/tracing.ts";
+import { workflowExecutionContext } from "#activities/temporal-context.ts";
+
+const COMPONENT = "agent-task";
+const DEFAULT_WORKFLOW_TYPE = "agentTaskWorkflow";
+
+export function jsonLog(
+  level: "info" | "warning" | "error",
+  message: string,
+  fields: Record<string, unknown> = {},
+): void {
+  const info = activityInfoOrUndefined();
+  const base: Record<string, unknown> = {
+    level,
+    msg: message,
+    component: COMPONENT,
+    ...getTraceContext(),
+    ...fields,
+  };
+  if (info !== undefined) {
+    Object.assign(base, info);
+  }
+  console.warn(JSON.stringify(base));
+}
+
+function activityInfoOrUndefined(): Record<string, unknown> | undefined {
+  try {
+    const info = Context.current().info;
+    return {
+      workflow: info.workflowType,
+      ...workflowExecutionContext(info),
+      activity: info.activityType,
+      attempt: info.attempt,
+    };
+  } catch {
+    return undefined;
+  }
+}
+
+export function captureWithContext(
+  error: unknown,
+  extra: Record<string, unknown> = {},
+): void {
+  Sentry.withScope((scope) => {
+    scope.setTag("component", COMPONENT);
+    const info = activityInfoOrUndefined();
+    if (info !== undefined) {
+      scope.setTag("workflow", String(info["workflow"]));
+      scope.setTag("activity", String(info["activity"]));
+    }
+    scope.setContext("agentTask", { ...info, ...extra });
+    Sentry.captureException(error);
+  });
+}
+
+export function safeHeartbeat(payload: Record<string, unknown>): void {
+  try {
+    Context.current().heartbeat(payload);
+  } catch {
+    // Local scripts can call activities directly; outside Temporal this is a no-op.
+  }
+}
+
+export function activityCancellationSignalOrUndefined():
+  | AbortSignal
+  | undefined {
+  try {
+    return Context.current().cancellationSignal;
+  } catch {
+    return undefined;
+  }
+}
+
+export function currentWorkflowType(): string {
+  try {
+    return Context.current().info.workflowType ?? DEFAULT_WORKFLOW_TYPE;
+  } catch {
+    return DEFAULT_WORKFLOW_TYPE;
+  }
+}
+
+export function startToCloseTimeoutMsOrUndefined(): number | undefined {
+  try {
+    return Context.current().info.startToCloseTimeoutMs;
+  } catch {
+    return undefined;
+  }
+}
+
+export function workflowId(): string {
+  try {
+    return (
+      Context.current().info.workflowExecution?.workflowId ??
+      `agent-task-local-${crypto.randomUUID()}`
+    );
+  } catch {
+    return `agent-task-local-${crypto.randomUUID()}`;
+  }
+}
